Type modifiers parsing support.
Currently, 'signed', 'unsigned', 'long' and 'short' keywords are supported as type modifiers.
The long and short implementations have been changed to be taken as type modifiers instead of an actual type.
Any type modifier without a type is defaulted to have type 'int' (aka a 'long' is a 'long int').
Type modifiers with real types ('float', 'double') and 'void' type are not allowed.