Refactored array type representation + Added indexing codegen
Array type now has a 'size' field instead of a 'count' node in declaration.
The 'size' is to be computed at compile time by evaluating the 'count' expression node.
Code generation is done by adding the offset using the 'index' node.
Pointers are dereferenced to get the array address, an array has an implicit address which doesn't require dereferencing.